Avoid clipping some photometric curves when displaying only relevant half spaces

+++ photometric_viewer/utils/coordinates.py
@@ -1,5 +1,5 @@
 from typing import Tuple
 
 
-def cartesian_to_screen(center: Tuple[int, int], point: Tuple[int, int]) -> Tuple[int, int]:
+def cartesian_to_screen(center: Tuple[float, float], point: Tuple[float, float]) -> Tuple[float, float]:
     return center[0] + point[0], center[1] - point[1]
